fix(eleventy-plugin-text-to-speech): avoid duplicated audio elements
When evaluating XPath expressions or CSS selectors in a HTML document,
we might find more than one match. Because of this, we need to discard
those matches that we already considered, otherwise for N matches we
would end up with NxN audio players in the HTML document.

The existing implementation couldn't address this issue.
This commit fixes it.